fix a bug

+++ PB17111577_宁雨亭_exp2/supervise/src/LogisticRegression.py
@@ -21,7 +21,7 @@ class LogisticRegression:
     
     def predict(self, x_test):
         y_pred = self._sigmoid(np.mat(x_test) * self.weight)
-        y_pred = list(map(lambda x: 1 if int(x[0]) > 0.5 else 0, y_pred))
+        y_pred = list(map(lambda x: 1 if float(x[0]) > 0.5 else 0, y_pred))
         return y_pred
